+++ src-tauri/src/lib.rs
@@ -67,6 +67,8 @@ pub fn run() {
         .plugin(tauri_plugin_store::Builder::default().build())
         .plugin(tauri_plugin_http::init())
         .plugin(tauri_plugin_notification::init())
+        .plugin(tauri_plugin_updater::Builder::default().build())
+        .plugin(tauri_plugin_process::init())
         .manage(AuthState::new(client_id, client_secret))
         .manage(TokenStore::new())
         .manage(GoogleClient::new())